requirements.txt
flask==3.1.2
mysql-connector-python==9.4.0
beautifulsoup4==4.13.5
requests==2.32.5
nltk==3.9.1
scikit-learn==1.7.2
numpy==2.3.2
